Move all operations on Fonts from System.Drawing.Common into a separate encapsulated class FontCache.
