+++ ClosedXML/ClosedXML.csproj
@@ -38,6 +38,9 @@
     </PackageReference>
     <PackageReference Include="Janitor.Fody" Version="1.8.0" PrivateAssets="all" />
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
+    <PackageReference Include="SixLabors.Fonts">
+      <Version>1.0.0-beta17</Version>
+    </PackageReference>
     <PackageReference Include="System.Drawing.Common" Version="4.5.0" />
   </ItemGroup>
