+++ ClosedXML/Extensions/FontBaseExtensions.cs
@@ -1,8 +1,6 @@
 // Keep this file CodeMaid organised and cleaned
 using ClosedXML.Utils;
 using System;
-using System.Collections.Generic;
-using System.Drawing;
 
 namespace ClosedXML.Excel
 {
@@ -23,45 +21,23 @@ namespace ClosedXML.Excel
             font.FontCharSet = sourceFont.FontCharSet;
         }
 
-        public static Double GetHeight(this IXLFontBase fontBase, Dictionary<IXLFontBase, Font> fontCache)
+        public static Double GetHeight(this IXLFontBase fontBase, FontCache fontCache)
         {
-            var font = GetCachedFont(fontBase, fontCache);
-            var textHeight = GraphicsUtils.MeasureString("X", font).Height;
+            var textHeight = fontCache.MeasureString("X", fontBase).Height;
             return (double)textHeight * 0.85;
         }
 
-        public static Double GetWidth(this IXLFontBase fontBase, String text, Dictionary<IXLFontBase, Font> fontCache)
+        public static Double GetWidth(this IXLFontBase fontBase, String text, FontCache fontCache)
         {
             if (String.IsNullOrWhiteSpace(text))
                 return 0;
 
-            var font = GetCachedFont(fontBase, fontCache);
-            var textWidth = GraphicsUtils.MeasureString(text, font).Width;
+            var textWidth = fontCache.MeasureString(text, fontBase).Width;
 
             double width = (textWidth / 7d * 256 - 128 / 7) / 256;
             width = Math.Round(width + 0.2, 2);
 
             return width;
-        }
-
-        private static Font GetCachedFont(IXLFontBase fontBase, Dictionary<IXLFontBase, Font> fontCache)
-        {
-            if (!fontCache.TryGetValue(fontBase, out Font font))
-            {
-                font = new Font(fontBase.FontName, (float)fontBase.FontSize, GetFontStyle(fontBase));
-                fontCache.Add(fontBase, font);
-            }
-            return font;
-        }
-
-        private static FontStyle GetFontStyle(IXLFontBase font)
-        {
-            FontStyle fontStyle = FontStyle.Regular;
-            if (font.Bold) fontStyle |= FontStyle.Bold;
-            if (font.Italic) fontStyle |= FontStyle.Italic;
-            if (font.Strikethrough) fontStyle |= FontStyle.Strikeout;
-            if (font.Underline != XLFontUnderlineValues.None) fontStyle |= FontStyle.Underline;
-            return fontStyle;
         }
     }
 }
