+++ ClosedXML/Utils/GraphicsUtils.cs
@@ -20,12 +20,5 @@ namespace ClosedXML.Utils
                 return threadLocalGraphics;
             }
         }
-
-        private static StringFormat defaultStringFormat = StringFormat.GenericTypographic;
-        public static SizeF MeasureString(string s, Font font)
-        {
-            SizeF result = Graphics.MeasureString(s, font, Int32.MaxValue, defaultStringFormat);
-            return result;
-        }
     }
 }
